refactor(hub): dissolve with-share, /bundle → deprecated alias + repoint importers
Replace the old src/with-share/bundle/index.ts barrel with a thin
deprecated alias at src/bundle/index.ts that re-exports the artifact
symbols from with-pod, the partition ops from with-cargo, and the shared
errors — so @noy-db/hub/bundle keeps resolving the identical surface.
Repoint the remaining cross-cutting importers (index.ts, noydb.ts,
vault.ts). src/with-share/ is now gone. Pure file moves plus import-path
rewrites; no behavior change.
