+++ packages/hub/src/vault.ts
@@ -26,9 +26,9 @@ import type { IndexDef } from './with-lookup/indexing/eager-indexes.js'
 import type { JoinableSource } from './query/index.js'
 import type { OnDirtyCallback } from './collection.js'
 import type { UnlockedKeyring, BundleRecipient } from './with-party/team/keyring.js'
-import { exportAccessibleData } from './with-share/bundle/export-accessible.js'
-import { withdrawAccessibleData } from './with-share/bundle/withdraw-accessible.js'
-import { requestWithdrawal, listWithdrawalRequests, approveWithdrawal, rejectWithdrawal } from './with-share/bundle/request-withdrawal.js'
+import { exportAccessibleData } from './with-audit/portability/export-accessible.js'
+import { withdrawAccessibleData } from './with-audit/portability/withdraw-accessible.js'
+import { requestWithdrawal, listWithdrawalRequests, approveWithdrawal, rejectWithdrawal } from './with-audit/portability/request-withdrawal.js'
 import type { MaterializedViewRegistry } from './with-formula/materialized-views/registry.js'
 import type { MaterializedViewStrategyHandle, MVQueryContext } from './with-formula/materialized-views/types.js'
 import type { OverlayedViewRegistry } from './with-formula/overlay-views/registry.js'
@@ -3423,7 +3423,7 @@ export class Vault {
     // import from bundle/* without forming a cycle. The dynamic
     // import is invoked once per fresh handle generation, which
     // is rare enough that the cost doesn't matter.
-    const { generateULID } = await import('./with-share/bundle/ulid.js')
+    const { generateULID } = await import('./with-pod/ulid.js')
     const handle = generateULID()
     const envelope: EncryptedEnvelope = {
       _noydb: NOYDB_FORMAT_VERSION,
